backend/requirements.txt
# LangGraph Agent Backend
# Uses Chinese LLMs via OpenRouter (same as frontend)

# Core
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0  # For OpenRouter compatibility
langsmith>=0.1.0  # Tracing

# Tools
tavily-python>=0.3.0  # Web search

# Utilities
httpx>=0.26.0
pydantic>=2.0.0
